fix: focus menu sidebar base on route
