+++ frontend/src/components/sidebar/SidebarItem.tsx
@@ -7,17 +7,15 @@ export type SidebarItemProp = {
 	title: string;
 	href: string;
 	isSelected?: boolean;
-	setActive?: (event: React.MouseEvent<HTMLAnchorElement>) => void;
 };
 
 export default function SidebarItem(props: SidebarItemProp) {
-	const { iconName, title, href, isSelected = false, setActive } = props;
+	const { iconName, title, href, isSelected = false } = props;
 	const [isHovered, setIsHovered] = useState(false);
 	return (
 		<Link
 			onMouseEnter={() => setIsHovered(true)}
 			onMouseLeave={() => setIsHovered(false)}
-			onClick={setActive}
 			className={classNames(
 				'flex w-52 py-2 px-5',
 				isSelected ? 'bg-zinc-100 rounded-r-3xl' : '',
